+++ package.json
@@ -19,7 +19,8 @@
     "build": "tsc -p .",
     "test": "jest",
     "coveralls": "jest --coverage && cat ./coverage/lcov.info | coveralls",
-    "lint": "cross-env-shell eslint './src/{**/*,*}.{js,ts}' './tests/{**/*,*}.{js,ts}'"
+    "lint": "cross-env-shell eslint './src/{**/*,*}.{js,ts}' './tests/{**/*,*}.{js,ts}'",
+    "generate-docs": "typedoc --entryPoints ./src --entryPointStrategy expand --out ./docs --plugin typedoc-theme-hierarchy --theme hierarchy"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^6.6.0",
@@ -36,6 +37,8 @@
     "typescript": "^5.2.2",
     "@types/ws": "^8.5.5",
     "ws": "^7.0.0",
-    "cross-env": "^7.0.3"
+    "cross-env": "^7.0.3",
+    "typedoc": "^0.25.1",
+    "typedoc-theme-hierarchy": "^4.1.1"
   }
 }
