+++ src/websocket_retry_options.ts
@@ -1,24 +1,8 @@
 import { Backoff } from "./backoff/backoff";
-import { WebsocketBuffer } from "./websocket_buffer";
-import { WebsocketEventListeners } from "./websocket_event";
 
-export interface WebsocketOptions {
-  /**
-   * The Buffer to use.
-   */
-  readonly buffer?: WebsocketBuffer;
-
-  /**
-   * The options for the connection-retry-strategy.
-   */
-  readonly retry?: WebsocketConnectionRetryOptions;
-
-  /**
-   * The initial listeners to add to the websocket.
-   */
-  readonly listeners?: WebsocketEventListeners;
-}
-
+/**
+ * Options for the websockets retry-strategy.
+ */
 export interface WebsocketConnectionRetryOptions {
   /**
    * The maximum number of retries before giving up. No limit if undefined.
